requirements/base.txt
djangorestframework
drf-yasg
python-environ
django-jazzmin
django-cors-headers
django-filters
djangorestframework-simplejwt